Use commit as image_tag
This makes for easier verification that the image used in a task
definition is the intended image. Also makes it easier to test locally
if there are issues in client applications. I.e. we can see which task
definition they use, which commit of learn-cli it corresponds to, and
check that out locally for testing.
I would prefer to use Tag, but it is not listed as one of the valid
template variables in the codeship docs.